// www/js/GameSearch/packages/discovery-common/src/ts/common/types/pageContext.ts
export enum PageContext {
  SearchPage = "searchPage",
  SortDetailPageDiscover = "sortDetailPageDiscover",
  SortDetailPageHome = "sortDetailPageHome",
  GameDetailPage = "gameDetailPage",
  GamesPage = "gamesPage",
  HomePage = "homePage",
  PeopleListInHomePage = "peopleListInHomePage",
  InterestCatcher = "interestCatcher",
  SearchLandingPage = "searchLandingPage",
  SpotlightPage = "spotlightPage",
  UserProfilePage = "userProfilePage",
  SongListPage = "songListPage",
}

export default {
  PageContext,
};